+++ NEWS.rst
@@ -6,7 +6,9 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Minor: Added ``set_mismatch_template()`` to configure the mismatch HTML template path.
+* Minor: Custom mismatch template paths now fail if they cannot be resolved.
+* Minor: Added template format validation for ``oldText`` and ``newText`` markers.
 
 2.0.0
 -----